+++ src/steps/index.test.ts
@@ -3,21 +3,15 @@ import {
   Recording,
 } from '@jupiterone/integration-sdk-testing';
 
-import { setupADORecording } from '../../test/recording';
+import { setupAzureRecording } from '../../test/recording';
+import { config } from '../../test/config';
+import { getMatchRequestsBy } from '../../test/getMatchRequestsBy';
 import { ADOIntegrationConfig } from '../types';
 import { fetchProjects } from './projects';
-import { fetchWorkitems } from './workitems';
+import { fetchWorkItems } from './workitems';
 import { fetchTeams } from './teams';
 import { fetchUsers } from './users';
 import { fetchAccountDetails } from './account';
-
-const DEFAULT_ORG_URL = 'https://dev.azure.com/tkcasey1';
-const DEFAULT_ACCESS_TOKEN = 'topsecret123'; //fake secret because we have a recording
-
-const integrationConfig: ADOIntegrationConfig = {
-  orgUrl: process.env.ORG_URL || DEFAULT_ORG_URL,
-  accessToken: process.env.ACCESS_TOKEN || DEFAULT_ACCESS_TOKEN,
-};
 
 jest.setTimeout(1000 * 60 * 1);
 
@@ -28,22 +22,25 @@ afterEach(async () => {
 });
 
 test('should collect data', async () => {
-  recording = setupADORecording({
+  recording = setupAzureRecording({
     directory: __dirname,
-    name: 'steps',
+    name: 'fullIntegrationTest',
+    options: {
+      matchRequestsBy: getMatchRequestsBy({ config: config }),
+    },
   });
 
   const context = createMockStepExecutionContext<ADOIntegrationConfig>({
-    instanceConfig: integrationConfig,
+    instanceConfig: config,
   });
 
   // Simulates dependency graph execution.
   // See https://github.com/JupiterOne/sdk/issues/262.
   await fetchAccountDetails(context);
   await fetchProjects(context);
-  await fetchWorkitems(context);
   await fetchUsers(context);
   await fetchTeams(context);
+  await fetchWorkItems(context);
 
   // Review snapshot, failure is a regression
   expect({